+++ src/main/java/org/sandag/abm/ctramp/SubtourDepartureAndDurationTime.java
@@ -71,7 +71,7 @@ public class SubtourDepartureAndDurationTime
     private boolean[]                       needToComputeLogsum;
     private double[]                        modeChoiceLogsums;
 
-    private int[]                           tempWindow;
+    private short[]                         tempWindow;
 
     // create an array to count the subtours propcessed within work tours
     // there are at most 2 work tours per person
@@ -195,7 +195,7 @@ public class SubtourDepartureAndDurationTime
         tourDepartureTimeChoiceSample = new int[numDepartureTimeChoiceAlternatives + 1];
         Arrays.fill(tourDepartureTimeChoiceSample, 1);
 
-        tempWindow = new int[modelStructure.getNumberOfTimePeriods() + 1];
+        tempWindow = new short[modelStructure.getNumberOfTimePeriods() + 1];
 
     }
 
@@ -223,7 +223,7 @@ public class SubtourDepartureAndDurationTime
             int numWorkTours = workTourList.size();
 
             // save a copy of this person's original time windows
-            int[] personWindow = person.getTimeWindows();
+            short[] personWindow = person.getTimeWindows();
             for (int w = 0; w < personWindow.length; w++)
                 tempWindow[w] = personWindow[w];
